+++ index.html
@@ -231,5 +231,6 @@
     <div class="img-upload__message  img-upload__message--loading">Загружаем...</div>
   </template>
 <script type="module" src="./js/main.js"></script>
+
 </body>
 </html>
